Survey screen a bug fixes
Adds a survey and a thank you page, with basic integration, survey results are saved in survey_results.csv evey sruvey done adds to it up.
now if pressed on change when adding a wrong categorie it will be automatically changend.  also i little guide for users has been added.

to save all as an .exe file is not working so far.